switch off screen state monitoring by default
Screen.Blanked monitoring is used to track possible transition to states with
lower/higher power cosumption. On the other hand POLLERR is received in the deep
sleep waking up device frequently. Fixing this issue needs a time to understand
how to do it, so screen state monitoring is switched of but can be controlled by STATEFS_UDEV_MONITOR_SCREEN environment variable

Signed-off-by: Denis Zalevskiy <denis.zalevskiy@jolla.com>