[power-udev] Derive OnBattery from charging state. Fixes MER#1818
The OnBattery property is supposed to be true when current is drawn from
battery and false when there is a charger that can provide enough power
to actually charge the battery. However, currently it is simply derived
from connected USB charger type - which leaves for example wireless
charging are stuck to OnBattery=true state.

Use charging state instead of usb charger state and set OnBattery when
the device is neither in charging nor in maintenance mode.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>